feat: improve rapyd-sdk error handling and logging
